requirements.txt
matplotlib
seaborn
tqdm
pandas
numpy
tensorflow<2.11
fastapi
uvicorn
python-multipart
pillow
